+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Bugfixes
-- Fixed a crash when the client doesn't have registrate on Fabric.
+- Fix EMI integration crashes due to an incorrect import.